pyproject.toml
[tool.poetry]
name = "django-startproject-template"
version = "0.1.0"
description = "Template para iniciar um novo projeto com a linguagem de programação Python e o framework web Django."
authors = ["Renato Cruz <zkpcvm6dz@mozmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.5"
django-environ = "^0.9.0"
gunicorn = "^20.1.0"
whitenoise = {extras = ["brotli"], version = "^6.2.0"}
uvicorn = {platform = 'linux', version = "^0.17.6"}
psycopg2-binary = "^2.9.3"

[tool.poetry.dev-dependencies]
pylint = "^2.14.3"
autopep8 = "^1.6.0"
django-debug-toolbar = "^3.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
